% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smfsb6.R
\name{gillespie}
\alias{gillespie}
\title{Gillespie algorithm for Stochastic Petri Net (SPN) (in R)}
\usage{
gillespie(N, n, ...)
}
\description{
Gillespie algorithm for Stochastic Petri Net (SPN) (in R)
}
\examples{
N = list()
N$M=c(x1=50,x2=100)
N$Pre=matrix(c(1,0,1,1,0,1),ncol=2,byrow=TRUE)
N$Post=matrix(c(2,0,0,2,0,0),ncol=2,byrow=TRUE)
N$h=function(x,t,th=c(th1=1,th2=0.005,th3=0.6)){
  with(as.list(c(x,th)),{
   return(c(th1*x1, th2*x1*x2, th3*x2 ))
  })
}


out = gillespie(N,10000)
op = par(mfrow=c(2,2))
plot(stepfun(out$t,out$x[,1]),pch="")
plot(stepfun(out$t,out$x[,2]),pch="")
plot(out$x,type="l")
par(op)
}
