% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smfsb6.R
\name{gillespie_C}
\alias{gillespie_C}
\title{Gillespie algorithm for Stochastic Petri Net (SPN) (in C)}
\usage{
gillespie_C(N, n, seed, ...)
}
\description{
Gillespie algorithm for Stochastic Petri Net (SPN) (in C)
}
\examples{
N = list()
N$M=c(x1=50,x2=100)
N$Pre=matrix(c(1,0,1,1,0,1),ncol=2,byrow=TRUE)
N$Post=matrix(c(2,0,0,2,0,0),ncol=2,byrow=TRUE)
N$h=function(x,t,th=c(th1=1,th2=0.005,th3=0.6)){
  with(as.list(th),{
    return(c(th1*x[1], th2*x[1]*x[2], th3*x[2] ))
  })
}

out <- gillespie_C(N = N,n = 1e4,seed = 1654)

matplot(out$t,out$x[-1,],type="l")
plot(out$x,type="l")
}
