% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smfsb5.R
\name{imdeath}
\alias{imdeath}
\title{Simulate Immigration-Death Process (in R)}
\usage{
imdeath(n = 20, x0 = 0, lambda = 1, mu = 0.1)
}
\description{
Simulate Immigration-Death Process (in R)
}
\examples{
plot(imdeath(n=30))
}
