% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smfsb5.R
\name{imdeath_C}
\alias{imdeath_C}
\title{Simulate Immigration-Death Process (in C)}
\usage{
imdeath_C(n = 20, x0 = 0, lambda = 1, mu = 0.1, seed)
}
\description{
Simulate Immigration-Death Process (in C)
}
\examples{
plot(imdeath_C(n=30,seed=42))
}
