% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smfsb5.R
\name{rcfmc}
\alias{rcfmc}
\title{Simulate CTMC (in R)}
\usage{
rcfmc(n, Q, pi0)
}
\description{
Simulate CTMC (in R)
}
\examples{
Q = matrix(c(-0.5,0.5,1,-1),ncol=2,byrow=TRUE)
pi0 = c(1,0)
plot(rcfmc(20,Q,pi0))
}
