% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smfsb5.R
\name{rdiff}
\alias{rdiff}
\title{Simulate Diffusion via Euler-Maruyama Approximation (in R)}
\usage{
rdiff(afun, bfun, x0 = 0, t = 50, dt = 0.01, ...)
}
\description{
Simulate Diffusion via Euler-Maruyama Approximation (in R)
}
\examples{
afun <- function(x,lambda,mu) {
 lambda-mu*x
}
bfun <- function(x,lambda,mu) {
 sqrt(lambda+mu*x)
 }
 plot(rdiff(afun,bfun,lambda=1,mu=0.1,t=30))
}
