% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smfsb5.R
\name{rdiff_C}
\alias{rdiff_C}
\title{Simulate Diffusion via Euler-Maruyama Approximation (in C)}
\usage{
rdiff_C(afun, bfun, x0 = 0, t = 50, dt = 0.01, seed, ...)
}
\description{
Simulate Diffusion via Euler-Maruyama Approximation (in C)
}
\examples{
afun <- function(x,lambda,mu) {
 lambda-mu*x
}
bfun <- function(x,lambda,mu) {
 sqrt(lambda+mu*x)
 }
 out <- rdiff_C(afun,bfun,seed = 42,lambda=1,mu=0.1)
 plot(ts(out,0.01))
}
