% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smfsb5.R
\name{rfmc}
\alias{rfmc}
\title{Simulate DTMC (in R)}
\usage{
rfmc(n, P, pi0)
}
\description{
Simulate DTMC (in R)
}
\examples{
P=matrix(c(0.9 ,0.1 ,0.2 ,0.8) ,ncol=2,byrow=TRUE)
pi0=c(0.5 ,0.5)
samplepath=rfmc (200 ,P , pi0 )
plot(samplepath)
}
