% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smfsb6.R
\name{simpleEuler}
\alias{simpleEuler}
\title{Euler Scheme for Numerical Integration of ODEs (in R)}
\usage{
simpleEuler(t = 50, dt = 0.001, fun, ic, ...)
}
\description{
Euler Scheme for Numerical Integration of ODEs (in R)
}
\examples{
lv <- function(x,t,k=c(k1=1,k2=0.1,k3=0.1)) {
  with(as.list(c(x,k)),{
   c( k1*x1 - k2*x1*x2 ,
     k2*x1*x2 - k3*x2 )
   })
 }
plot(simpleEuler(t=100,fun=lv,ic=c(x1=4,x2=10)))
}
