% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smfsb6.R
\name{simpleEuler_C}
\alias{simpleEuler_C}
\title{Euler Scheme for Numerical Integration of ODEs (in C)}
\usage{
simpleEuler_C(t = 50, dt = 0.001, fun, ic, ...)
}
\description{
Euler Scheme for Numerical Integration of ODEs (in C)
}
\examples{
lv <- function(x,t,k=c(k1=1,k2=0.1,k3=0.1)) {
  with(as.list(c(x,k)),{
    c( k1*x[1] - k2*x[1]*x[2] ,
       k2*x[1]*x[2] - k3*x[2] )
   })
 }
plot(simpleEuler_C(t=100,fun=lv,ic=c(4,10)))
}
